// src/types/user.ts
export interface User {
  userId: string;
  userName: string;
  email: string;
  phoneNumber: string;
  bvn: string;
  gender: string;
  maritalStatus: string;
  children: string;
  residenceType: string;
  education: string;
  employmentStatus: string;
  employmentSector: string;
  employmentDuration: string;
  officeEmail: string;
  monthlyIncome: string;
  loanRepayment: string;
  twitter: string;
  facebook: string;
  instagram: string;
  guarantorFullName: string;
  guarantorPhone: string;
  guarantorEmail: string;
  guarantorRelationship: string;
  password: string;
}